requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
charset-normalizer==2.1.1
click==8.1.3
discord.py==2.1.0
dnspython==2.2.1
fastapi==0.88.0
frozenlist==1.3.3
graphql-core==3.2.3
h11==0.14.0
httptools==0.5.0
idna==3.4
motor==3.1.1
multidict==6.0.3
pydantic==1.10.2
pymongo==4.3.3
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.5
PyYAML==6.0
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
strawberry-graphql==0.151.0
typing_extensions==4.4.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
yarl==1.8.2
